Disambiguate float and double sqrt() calls -- perhaps in poor style. Would a lookup table make sense here? It depends on the range of m_order.
svn path=/trunk/KDE/kdegames/ksudoku/; revision=1062381

+++ src/logic/skgraph.cpp
@@ -76,7 +76,7 @@ inline void SKGraph::addConnection(int i, int j)
 
 void ksudoku::GraphSudoku::init()
 {
-	int base = static_cast<int>(sqrt(m_order));
+	int base = static_cast<int>(sqrt((float)m_order));
 	
 	m_ruleset = new Ruleset();
 	ItemBoard *board = new ItemBoard(m_order, m_order);
@@ -146,7 +146,7 @@ void ksudoku::GraphSudoku::init()
 
 void ksudoku::GraphRoxdoku::init()
 {
-	int base = static_cast<int>(sqrt(m_order));
+	int base = static_cast<int>(sqrt((float)m_order));
 	
 	m_ruleset = new Ruleset();
 	ItemBoard *board = new ItemBoard(base, base, base);
